Make teleport detection more lenient
If you fall from the center of The Tower of Antumbra, when you transition from the 2nd to last to 3rd to last screen of the Halted Ruin (you will be along the right side of the screen) your Y velocity can exceed the max fall for a frame- probably related to moving into collision/a slope when moving to the 3rd to last screen- I've seen Y velocities as high as about 16.5 which was tripping teleport detection.

Bumping the multiplier for teleport detection to avoid false positives in this situations (and probably similar).